#ALSA-14 #ALSA-25 generate crud API routes with apiResource
exposed methods:
- index
- store
- show
- update
- destroy
